requirements.txt
async-timeout==3.0.1
discord.py==1.2.3
enum34==1.1.6
multidict==4.5.2
netifaces==0.10.4
numpy==1.16.2
oauthlib==3.0.1
pipenv==11.9.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pylint==2.3.1
pymacaroons==0.13.0
pymongo==3.8.0
PyNaCl==1.3.0
pytz==2019.1
PyYAML==5.1
reportlab==3.5.18
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
urllib3==1.24.2

